requirement.txt
json
requests
mysql-connector-python
MySQL-python
